requirements.txt
google-genai
langchain
langchain-core
python-dotenv